resolve conflicts2
